Move console to the windows folder
