+++ trview.app/Windows/Console/Console.cpp
