+++ trview.app/Windows/Console/Console.h
